ECA-1272: Improved authorization-related error message when editing an already existing user.
git-svn-id: https://svn.cesecore.eu/svn/ejbca-restricted/trunk/ejbca@7607 37aeb1de-b14b-0410-9a0e-82b28170d228

+++ src/java/org/ejbca/core/protocol/ws/client/StressTestCommand.java
@@ -268,7 +268,7 @@ public class StressTestCommand extends EJBCAWSRABaseCommand implements IAdminCom
         public boolean doIt() throws Exception {
             if ( this.doCreateNewUser ) {
                 this.jobData.passWord = "foo123";
-                this.jobData.userName = "WSTESTUSER"+StressTestCommand.this.performanceTest.getRandom().nextInt();
+                this.jobData.userName = "WSTESTUSER"+StressTestCommand.this.performanceTest.getRandom().nextLong();
             }
             this.user.setSubjectDN(this.jobData.getDN());
             this.user.setUsername(this.jobData.userName);
